__init__.py files should not contain significant code.
