+++ simtfl/bft/chain.py
@@ -1,13 +1,5 @@
 """
 Abstractions for Byzantine Fault-Tolerant protocols.
-
-The model of a BFT protocol assumed here was written for the purpose
-of Crosslink, based on the example of adapted-Streamlet ([CS2020] as
-modified in [Crosslink]). It might not be sufficient for other BFT
-protocols — but that's okay; it's a prototype.
-
-[CS2020] https://eprint.iacr.org/2020/088.pdf
-[Crosslink] https://hackmd.io/JqENg--qSmyqRt_RqY7Whw?view
 """
